green tests for rds and ec2
