+++ lib/ex_aws/operation/rest_query.ex
@@ -1,0 +1,8 @@
+defmodule ExAws.Operation.RestQuery do
+  defstruct [
+    http_method: nil,
+    path: "/",
+    params: %{},
+    service: nil,
+  ]
+end
